build functions

+++ promote_from_staging.sh
@@ -95,14 +95,22 @@ git config --global push.default simple
 git checkout master
 git pull
 git merge ${CIRCLE_BRANCH} -m "Auto merge"
+
 echo ${RELEASE} > .release
 echo ${RELEASE} ${TAG} ${CODENAME} ${CIRCLE_SHA1} > .release.details
-envsubst '${RELEASE}:${BLURB}:${FOOTER}:${HEADER}:${STATE_SHELVED}:${STATE_EXPERIMENTAL}:${STATE_ACTIVE}:${STATE_PRE_ALPHA}:${STATE_ALPHA}:${STATE_BETA}:${STATE_PROD}:${TUTUM}' < README.tmpl.md > README.md
+
+if [[ -f README.md ]]
+then
+    envsubst '${RELEASE}:${BLURB}:${FOOTER}:${HEADER}:${STATE_SHELVED}:${STATE_EXPERIMENTAL}:${STATE_ACTIVE}:${STATE_PRE_ALPHA}:${STATE_ALPHA}:${STATE_BETA}:${STATE_PROD}:${TUTUM}' < README.tmpl.md > README.md
+    git add README.md
+fi
+
 if [ -f tutum.tmpl.yml ]
 then
     envsubst '${RELEASE}' < tutum.tmpl.yml > tutum.yml
 fi
-git add README.md .release .release.details
+git add .release .release.details
+
 git commit -a -m "Promotion of ${RELEASE}" || :
 git fetch
 git push
